CRISPRofftargetHunter

Software designed for efficient, precise and fast identification of all off-targets for given CRISPR guideRNA's. It leverages couple of alghoritms designed specifically for this purpose.

CRISPRofftargetHunter is extensively tested:

unit tests for each function

friction tests where three different implementations of the same functionality must report the same results

end-to-end tests where we run whole pipeline on specially designed sample genome and compare results with CRISPRitz software

CRISPRofftargetHunter is designed specifically for CRISPR alignments of guideRNA's, allowing for deletions, insretions and mismatches. Implemented alghoritms allow you to find off-targets within distance as large as you want!

CRISPRofftargetHunter has an alghoritm (see: build_sketchDB and search_sketchDB) designed for super fast estimation of number of off-targets in the genome. These estimations can never report counts less than reality, but can only over-estimate! This can be used to quickly design libraries of guideRNA's for the entire genomes, as promising guides can be quickly sorted using CRISPRofftargetHunter.

CRISPRofftargetHunter has support for multiple-cores, we use standard julia configuration for that.

Citation

TODO

LICENSE

Copyright (C) 2021 Kornel Labun

This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.

You should have received a copy of the GNU Affero General Public License along with this program. If not, see https://www.gnu.org/licenses/.

Self-contained build

It is possible to build CRISPRofftargetHunter into standalone application. This can be achieved by running ./build_standalone.sh script from the main directory. Script will produce binary in a new folder outside the main directory. Then you can run from inside that folder ./bin/CRISPRofftargetHunter --help or ./bin/CRISPRofftargetHunter build --help.

Main API

Run CRISPRofftargetHunter as an application. From the directory of the package run:

julia --threads 4 --project="." CRISPRofftargetHunter.jl --help

Example

TODO

Public Interface

You can also use CRISPRofftargetHunter as a normal julia package with the exported functions.

CRISPRofftargetHunter.Motif — Type

Motif( alias::String, fwdmotif::String, fwdpam::String, forward_strand::Bool = true, reverse_strand::Bool = true, distance::Int = 4, extends5::Bool = true)

Motif(alias::String)

Motif defines what we search on the genome, what can be identified as an off-target.

Arguments

alias - alias of the motif for easier identification e.g. Cas9

fwdmotif - Motif that indicates where is PAM inside fwdpam. For example for Cas9 it is 20*N + XXX: NNNNNNNNNNNNNNNNNNNNXXX

fwdpam - Motif in 5'-3' that will be matched on the reference (without the X). For example for Cas9 it is 20*X + NGG: XXXXXXXXXXXXXXXXXXXXNGG

forward - If false will not match to the forward reference strand.

reverse - If false will not match to the reverse reference strand.

distance - How many extra nucleotides are needed for a search? This will indicate within what distance we can search for off-targets. When we don't have those bases we use DNA_Gap.

extend5 - Defines how off-targets will be aligned to the guides and where extra nucleotides will be added for alignment within distance. Whether to extend in the 5' and 3' direction. Cas9 is extend5 = true.

Example for Cas9 where we want to search for off-targets within distance of 4:
  alias:    Cas9
  fwdmotif: NNNNNNNNNNNNNNNNXXX
  fwdpam:   XXXXXXXXXXXXXXXXNGG
  forward:  true
  reverse:  true
  distance: 4
  extend5:  true

Alignments will be performed from opposite to the extension direction (which is deifned by extend5).

Examples

Motif('Cas9')
Motif('Cas9', 'NNNNNNNNNNNNNNNNXXX', 'XXXXXXXXXXXXXXXXNGG'. true, true, 3, true)

CRISPRofftargetHunter.build_linearDB — Function

build_linearDB( name::String, genomepath::String, motif::Motif, storagedir::String, prefix_len::Int = 7)

Build a DB of offtargets for the given motif, DB groups off-targets by their prefixes.

Will return a path to the database location, same as storagedir. When this database is used for the guide off-target scan it is similar to linear in performance, hence the name.

There is an optimization that if the alignment becomes imposible against the prefix we don't search the off-targets grouped inside the prefix. Therefore it is advantageous to select larger prefix than maximum search distance, however in that case number of files also grows.

CRISPRofftargetHunter.search_linearDB — Function

search_linearDB(storagedir::String, guides::Vector{LongDNA{4}}, dist::Int = 4; detail::String = "")

Will search the previously build database for the off-targets of the guides. Assumes your guides do not contain PAM, and are all in the same direction as you would order from the lab e.g.:

` 5' - ...ACGTCATCG NGG - 3' -> will be input: ...ACGTCATCG guide PAM

3' - CCN GGGCATGCT... - 5' -> will be input: ...AGCATGCCC PAM guide `

Arguments

dist - defines maximum levenshtein distance (insertions, deletions, mismatches) for which off-targets are considered. detail - path and name for the output file. This search will create intermediate files which will have same name as detail, but with a sequence prefix. Final file will contain all those intermediate files. Leave detail empty if you are only interested in off-target counts returned by the linearDB.

CRISPRofftargetHunter.build_sketchDB — Function

build_sketchDB( name::String, genomepath::String, motif::Motif, storagedir::String, probability_of_error::Float64 = 0.001; max_count::Int = 255)

Build Count-Min-Sketch database that can never under-estimate the counts of off-targets.

Arguments

name - Your name for this database. genomepath - Path to the genome in either .2bit or .fa format. motif- See Motif type. storagedir- Where should be the output folder. probability_of_error- This is error level for each individual off-target search. This error will be propagated when doing searches for larger distances and therefore will not be perfectly reflected in the results.

max_count - What is the maximum count for each specific off-target in the genome. We normally care only for guides that have low off-target count, therefore we can keep this value also low, this saves space and memory usage.

CRISPRofftargetHunter.search_sketchDB — Function

search_sketchDB( storagedir::String, guides::Vector{LongDNA{4}}, dist::Int = 2)

For each of the input guides will check in the sketchDB what value Count-Min Sketch has for it in the first, and in the second column will output kmer sum. Results are estimations of offtarget counts in the genome.

If CMS column is 0, it is guaranteed this guide has no 0-distance off-targets in the genome!

Also, maximum count for each off-target in the database is capped, to max_count specified during building of sketchDB. This means that counts larger than max_count (default 255) are no longer estimating correclty, but we know at least 255 off-targets exist in the genome. Its likely you would not care for those guides anyhow.

CRISPRofftargetHunter.build_treeDB — Function

build_treeDB( name::String, genomepath::String, motif::Motif, storagedir::String, prefix_len::Int = 7)

Build a Vantage Point tree DB of offtargets for the given motif, DB groups off-targets by their prefixes, each prefix has its own Vantage Point tree.

Will return a path to the database location, same as storagedir.

There is an optimization that if the alignment becomes imposible against the prefix we don't search the off-targets grouped inside the prefix. Therefore it is advantageous to select larger prefix than maximum search distance, however in that case number of files also grows.

CRISPRofftargetHunter.search_treeDB — Function

search_treeDB(storagedir::String, guides::Vector{LongDNA{4}}, dist::Int = 4; detail::String = "")

Will search the previously build database for the off-targets of the guides. Assumes your guides do not contain PAM, and are all in the same direction as you would order from the lab e.g.:

5' - ...ACGTCATCG NGG - 3'  -> will be input: ...ACGTCATCG
     guide        PAM

3' - CCN GGGCATGCT... - 5'  -> will be input: ...AGCATGCCC
     PAM guide

Arguments

dist - defines maximum levenshtein distance (insertions, deletions, mismatches) for which off-targets are considered.

detail - path and name for the output file. This search will create intermediate files which will have same name as detail, but with a sequence prefix. Final file will contain all those intermediate files. Leave detail empty if you are only interested in off-target counts returned by the searchDB.

CRISPRofftargetHunter.inspect_treeDB — Function

inspect_treeDB(storagedir::String; levels::Int = 5, inspect_prefix::String = "")

See small part of the full vantage point tree of the treeDB.