Public Interface

When using ARTEMIS as a normal julia package with the exported functions we provide many methods for working with CRISPR gRNA and off-targets for your convenience.

Abstract gRNA

To encapsulate motifs that we search for on the genome, we can use abstract definition of gRNA/off-target, which is defined in Motif data structure.

ARTEMIS.Motif — Type

Motif(
    alias::String, 
    fwdmotif::String, 
    fwdpam::String, 
    forward_strand::Bool = true, 
    reverse_strand::Bool = true, 
    distance::Int = 4, 
    extends5::Bool = true,
    ambig_max::Int = 5)

Motif(alias::String)

Motif defines what we search on the genome, what can be identified as an off-target.

Arguments

alias - alias of the motif for easier identification e.g. Cas9

fwdmotif - Motif that indicates where is PAM inside fwdpam. For example for Cas9 it is 20*N + XXX: NNNNNNNNNNNNNNNNNNNNXXX

fwdpam - Motif in 5'-3' that will be matched on the reference (without the X). For example for Cas9 it is 20*X + NGG: XXXXXXXXXXXXXXXXXXXXNGG

forward - If false will not match to the forward reference strand.

reverse - If false will not match to the reverse reference strand.

distance - How many extra nucleotides are needed for a search? This will indicate within what distance we can search for off-targets. When we don't have those bases we use DNA_Gap.

extend5 - Defines how off-targets will be aligned to the guides and where extra nucleotides will be added for alignment within distance. Whether to extend in the 5' and 3' direction. Cas9 is extend5 = true. ambig_max- How many ambiguous bases are allowed in the pattern?

Example for Cas9 where we want to search for off-targets within distance of 4:
  alias:    Cas9
  fwdmotif: NNNNNNNNNNNNNNNNNNNNXXX
  fwdpam:   XXXXXXXXXXXXXXXXXXXXNGG
  forward:  true
  reverse:  true
  distance: 4
  extend5:  true
  ambig_max:5 

Alignments will be performed from opposite to the extension direction (which is deifned by extend5).

Examples

julia> Motif("Cas9")
Motif("Cas9", NNNNNNNNNNNNNNNNNNNNNGG, CCNNNNNNNNNNNNNNNNNNNNN, 21:23, 1:3, 3, true, 0)

julia> Motif("test name", "NNNNNNNNNNNNNNNNNNNNXXX", "XXXXXXXXXXXXXXXXXXXXNGG", true, true, 3, true, 5)
Motif("test name", NNNNNNNNNNNNNNNNNNNNNGG, CCNNNNNNNNNNNNNNNNNNNNN, 21:23, 1:3, 3, true, 5)

ARTEMIS.length_noPAM — Function

length_noPAM(motif::Motif)

Calculate what is the length of the motif, with extension, but without PAM. Effectively, size of the off-target on the genome.

Examples

julia> length_noPAM(Motif("Cas9"))
20

ARTEMIS.setdist — Function

setdist(motif::Motif, distance::Int)

Set the distance (maximum value of allowed mismatches, deletion, insertions) that are allowed during alignment.

Examples

julia> setdist(Motif("Cas9"), 15)
Motif("Cas9", NNNNNNNNNNNNNNNNNNNNNGG, CCNNNNNNNNNNNNNNNNNNNNN, 21:23, 1:3, 15, true, 0)

ARTEMIS.setambig — Function

setambig(motif::Motif, ambig::Int)

Set the ambiguity (how many ambigous bases are allowed) level for motif.

Examples

julia> setambig(Motif("Cas9"), 15)
Motif("Cas9", NNNNNNNNNNNNNNNNNNNNNGG, CCNNNNNNNNNNNNNNNNNNNNN, 21:23, 1:3, 3, true, 15)

Find off-targets

ARTEMIS.DBInfo — Type

DBInfo(filepath::String, name::String, motif::Motif; vcf_filepath::String = "")

Motif defines what genome file is being used for searches.

Arguments

filepath - Path to the genome file, if file is fasta (ends with .fa or .fasta or .fna) make sure you also have fasta index file with extension .fai. Alternatively, you can use .2bit genome file.

name - Your name for this instance of DBInfo: the genome with connection to the motif and vcf file.

motif - Motif object defining search parameters

vcf_filepath - Optional. Path to the VCF file to include in the searches.

Alignments will be performed from opposite to the extension direction (which is deifned by extend5).

Examples

# use ARTEMIS example genome
genome = joinpath(
    vcat(
        splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
        "test", "sample_data", "genome", "semirandom.fa"))
# construct example DBInfo
DBInfo(genome, "Cas9_semirandom_noVCF", Motif("Cas9"))

ARTEMIS.gatherofftargets! — Function

gatherofftargets!(output::T, dbi::DBInfo) 
    where {T<:Union{Vector{String}, Vector{UInt64}, Vector{UInt128}}}

Gathers all off-targets that conform to the given dbi Motif.

This function appends to the output during the run, however it will also return all ambiguous guides in return object. We use UInt64 and UInt128 to compress space that the gRNAs use. When using large genomes or non-specific PAMs you might run out of memory when using this function.

Examples

# use ARTEMIS example genome
genome = joinpath(
    vcat(
        splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
        "test", "sample_data", "genome", "semirandom.fa"))
# construct example DBInfo
dbi = DBInfo(genome, "Cas9_semirandom_noVCF", Motif("Cas9"))
# finally gather all off-targets
guides = Vector{String}()
ambig = gatherofftargets!(guides, dbi)

Align gRNA and off-target

ARTEMIS.isinclusive — Function

isinclusive(x::S, y::S) where {S<:BioSymbol}

Check whether x contains all options of y or y contains all options of x. Can be used as replacement in place of BioSequences.iscompatible or BioSequences.isequal.

Examples

julia> isinclusive(DNA_N, DNA_A)
true

julia> isinclusive(DNA_W, DNA_R) # W is (A/T), R is (A/G)
false

julia> isinclusive(DNA_W, DNA_A)
true

ARTEMIS.hamming — Function

hamming(s1::T, s2::K, ismatch = iscompatible) where {T <: BioSequence, K <: BioSequence}

Check Hamming distance between s1 and s2 using as matching definition ismatch. Should be faster than pairalign(HammingDistance(), s1, s2). Make sure inputs are of the same length.

Examples

julia> hamming(dna"ACGC", dna"AWRC")
1

ARTEMIS.levenshtein — Function

levenshtein(guide::T, ref::K, k::Int = 4,
    ismatch::Function = iscompatible) where {T <: BioSequence, K <: BioSequence}

Calculate Levenshtein distance bounded by k maximum edit distance with a twist for guide + off-target comparisons.

Levenshtein distance is the minimum number of operations (consisting of insertions, deletions, substitutions of a single character) required to change one string into the other. guide input seqeunce is a gRNA sequence, ref input is reference sequence with expansion on the 3' end of k bases. This extension will not count toward the score, if it is not covered with aligned guide. Return k + 1, if distance is higher than k and terminate early. This function should be 10x faster than pairalign(LevenshteinDistance(), guide, ref, distance_only = true).

Notice that guide and ref have to be oriented towards left side e.g. PAM-guide, and PAM-offtarget-extension! Take a look at the examples below to understand why.

Examples

julia> levenshtein(dna"ATGA", dna"AGACCT") # CCT as extension of the ref does not count towards the score in optimal alignment
1

julia> levenshtein(dna"ATGATCG", dna"AGAAATCGATG") # ATG does not count in optimal alignment ATG--ATCG/A-GAAATCG
3

ARTEMIS.Aln — Type

struct Aln
    guide::String
    ref::String
    dist::Int
end

Simple data structure to hold information on the optimal alignment. Therefore guide and ref may contain DNA_Gap as these are aligned sequences. Function align returns this object.

ARTEMIS.align — Function

align(guide::T, ref::K, k::Int = 4,
    ismatch::Function = iscompatible) where {T <: BioSequence, K <: BioSequence}

Calculate Levenshtein distance bounded by k maximum edit distance with a twist for guide + off-target comparisons, but return also the alignment as an Align object.

Levenshtein distance is the minimum number of operations (consisting of insertions, deletions, substitutions of a single character) required to change one string into the other. guide input seqeunce is a gRNA sequence, ref input is reference sequence with expansion on the 3' end of k bases. This extension will not count toward the score, if it is not covered with aligned guide. Return k + 1, if distance is higher than k and terminate early.

Notice that guide and ref have to be oriented towards left side e.g. PAM-guide, and PAM-offtarget-extension! Take a look at the examples below to understand why.

Examples

julia> align(dna"ATGA", dna"AGACCT") # CCT as extension of the ref does not count towards the score in optimal alignment
Aln("ATGA", "A-GA", 1)

julia> align(dna"ATGATCG", dna"AGAAATCGATG") # ATG does not count
Aln("ATG--ATCG", "A-GAAATCG", 3)

Utils

gRNAs and kmers

ARTEMIS.getseq — Function

getseq(n = 20, letters = ['A', 'C', 'G', 'T'])

Randomize sequence of length n from letters.

ARTEMIS.as_kmers — Function

as_kmers(x::LongDNA{4}, kmer_size::Int)

Transforms x into vector of kmers of size kmer_size. All ambiguous bases will be expanded.

Examples

julia> as_kmers(dna"ACTGG", 4)
2-element Vector{LongSequence{DNAAlphabet{4}}}:
 ACTG
 CTGG

ARTEMIS.as_skipkmers — Function

as_skipkmers(x::LongDNA{4}, kmer_size::Int)

Transforms x into vector of skip-kmers of size kmer_size. All ambiguous bases will be expanded. Leftover-bases are ignored!

Examples

julia> as_skipkmers(dna"ACTGG", 2)
2-element Vector{LongSequence{DNAAlphabet{4}}}:
 AC
 TG

ARTEMIS.all_kmers — Function

all_kmers(size = 4; alphabet = [DNA_A, DNA_C, DNA_G, DNA_T]

Make a list of all possible kmers with givensize using bases in the alphabet.

Examples

julia> all_kmers(2; alphabet = [DNA_A, DNA_N])
4-element Vector{LongSequence{DNAAlphabet{4}}}:
 AA
 AN
 NA
 NN

ARTEMIS.minkmersize — Function

minkmersize(len::Int = 20, d::Int = 4)

Pidgeon hole principle: minimum k-mer size that is required for two strings of size len to be aligned within distance of d.

Examples

julia> minkmersize(20, 3)
5

julia> minkmersize(20, 6)
2

Persistence

ARTEMIS.save — Function

save( object::Any, destination::String)

Uses julia serializer to save the data to binary format. Read more about serialization. Notice that:

This function will overwrite destination!

This serialization is dependent on julia build! This means files can fail to work when reloaded across different julia builds.

ARTEMIS.load — Function

load(destination::String)

Load file saved with save function. This may not load properly files saved in other julia builds.

Alignment-free filters for gRNAs

ARTEMIS.build_hashDB — Function

build_hashDB(
    name::String,
    genomepath::String, 
    motif::Motif,
    storage_dir::String;
    seed::UInt64 = UInt64(0x726b2b9d438b9d4d),
    max_iterations::Int = 10,
    max_count::Int = 10,
    precision::DataType = UInt16)

Prepare hashDB index for future searches using search_hashDB.

Arguments

name - Your prefered name for this index for easier identification.

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

motif - Motif deines what kind of gRNA to search for.

storage_dir - Folder path to the where index will be saved with name hashDB.bin.

seed - Optional. Seed is used during hashing for randomization.

max_iterations - When finding hashing structure for binary fuse filter it might fail sometimes, we will retry max_iterations number of times though.

max_count - Above this count we put all unique off-targets into one bin. Put number here that is the minimum number of off-targets that you think is fine for the distance of 1.

precision- The higher the precision the larger the database, but also chances for error decrease dramatically. We support UInt8, UInt16, and UInt32.

Examples

# make a temporary directory
tdir = tempname()
hdb_path = joinpath(tdir, "hashDB")
mkpath(hdb_path)

# use ARTEMIS example genome
genome = joinpath(
    vcat(
        splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
        "test", "sample_data", "genome", "semirandom.fa"))

# finally, build a hashDB
build_hashDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 1, ambig_max = 0), 
    hdb_path)

ARTEMIS.search_hashDB — Function

search_hashDB(
    storage_dir::String,
    guides::Vector{LongDNA{4}},
    right::Bool)

Estimate off-target counts for guides using hashDB stored at storage_dir.

Probabilistic filter offers a guarantee that it will always be correct when a sequence is in the set (no false negatives), but may overestimate that a sequence is in the set while it is not (false positive) with low probability. If both columns in the results are 0, it is guaranteed this gRNA has no off-targets in the genome!

Also, maximum count for each off-target in the database is capped, to max_count specified during building of hashDB. This means that counts larger than max_count are no longer estimating correctly. Its likely you would not care for those guides anyhow.

right argument specifies whether the databse should be checked in direction from unique off-targets which occur once to increasingly more occuring off-targets up until max_count is reached, which may result in assuming lower than real off-target counts (underestimate) for some of the sequences, however this approach will not reject any gRNAs that should not be rejected and is suitable for filtering of gRNAs we do not need. Left (right = false, or hight-counts to low-counts) approach is also supported, which can be used for ordering of gRNAs to the best of database ability. Left approach may overestimate counts for some gRNAs. When gRNA is reported as off-target free it is also guaranteed to be true in both cases (low-to-high and high-to-low).

Examples

# prepare libs
using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
hdb_path = joinpath(tdir, "hashDB")
mkpath(hdb_path)

# use ARTEMIS example genome
coh_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(coh_path, "test", "sample_data", "genome", "semirandom.fa"))

# build a hashDB
build_hashDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 1, ambig_max = 0), 
    hdb_path)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(coh_path, "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)

# finally, get results!
hdb_res = search_hashDB(hdb_path, guides, false)

Find all off-targets

Prefix-Suffix partial alignment

ARTEMIS.build_linearDB — Function

build_linearDB(
    name::String,
    genomepath::String,
    motif::Motif,
    storage_dir::String,
    prefix_len::Int = 7)

Prepare linearDB index for future searches using search_linearDB.

Will return a path to the database location, same as storage_dir. When this database is used for the guide off-target scan it is similar to linear in performance, hence the name. There is an optimization that if the alignment becomes imposible against the prefix we don't search the off-targets grouped inside the prefix. Therefore, it is advantageous to select much larger prefix than maximum search distance, however in that case number of files also grows. For example, if interested with searches within distance 4, preferably use prefix length of 7 or 8.

Arguments

name - Your prefered name for this index for easier identification.

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

motif - Motif deines what kind of gRNA to search for.

storage_dir - Folder path to the where index will be saved with name linearDB.bin and many prefix files.

prefix_len - Size of the prefix by which off-targets are indexed. Prefix of 8 or larger will be the fastest, however it will also result in large number of files.

Examples

# make a temporary directory
tdir = tempname()
ldb_path = joinpath(tdir, "linearDB")
mkpath(ldb_path)

# use ARTEMIS example genome
genome = joinpath(
    vcat(
        splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
        "test", "sample_data", "genome", "semirandom.fa"))

# finally, build a linearDB
build_linearDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 1, ambig_max = 0), 
    ldb_path)

ARTEMIS.search_linearDB — Function

search_linearDB(
    storage_dir::String, 
    guides::Vector{LongDNA{4}}, 
    dist::Int = 4; 
    detail::String = "")

Find all off-targets for guides within distance of dist using linearDB located at storage_dir.

Assumes your guides do not contain PAM, and are all in the same direction as you would order from the lab e.g.:

5' - ...ACGTCATCG NGG - 3'  -> will be input: ...ACGTCATCG
     guide        PAM
    
3' - CCN GGGCATGCT... - 5'  -> will be input: ...AGCATGCCC
     PAM guide

Arguments

dist - Defines maximum levenshtein distance (insertions, deletions, mismatches) for which off-targets are considered.

detail - Path and name for the output file. This search will create intermediate files which will have same name as detail, but with a sequence prefix. Final file will contain all those intermediate files. Leave detail empty if you are only interested in off-target counts returned by the linearDB.

Examples

# make a temporary directory
tdir = tempname()
ldb_path = joinpath(tdir, "linearDB")
mkpath(ldb_path)

# use ARTEMIS example genome
coh_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(
    vcat(
        coh_path, 
        "test", "sample_data", "genome", "semirandom.fa"))

# build a linearDB
build_linearDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 3, ambig_max = 0), 
    ldb_path)

# load up example gRNAs
using BioSequences
guides_s = Set(readlines(joinpath(vcat(coh_path, "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, get results!
ldb_res = search_linearDB(ldb_path, guides, 3)

ARTEMIS.build_motifDB — Function

build_motifDB(
    name::String,
    genomepath::String,
    motif::Motif,
    storage_dir::String,
    prefix_len::Int = 7;
    skipmer_size::Int = Int(floor(length_noPAM(motif) / (motif.distance + 3))))

Prepare motifDB index for future searches using search_motifDB.

Will return a path to the database location, same as storage_dir. When this database is used for the guide off-target scan it is similar to search_linearDB, however additional filter is applied on top of prefix filtering. Suffixes are used for next filter, similarly to pidgeon hole principle - depending on the size of the skipkmer skipkmer_size. For example, Cas9 off-target within distance 4 (d) might be 20bp long. We skip prefix_len of 7, and are left with 13bp which can be split into 3 skipmers (r) of size 4, 1bp will be left unused. However when searching within distance of 4 and for prefix where initial alignment was of distance 3 (m) and adjustment paramter is 0 (a). We are obliged to find at least k - (d - m + a) which is 3 - (4 - 3 + 0) = 2 this many skimpers inside the off-targets.

There exist also another approach which builds on the idea that it might be more efficient to find at least two kmers of smaller size (named 01*0 seed) rather than one larger kmer (pidgeon hole principle). You can use the adjust option for that during search_linearDB step.

Be sure to understand implications of using motifDB as using wrong parameters on skipmer_size might result in leaky filtering in relation to the asumed distance dist and adjustment adjust during search step in search_motifDB.

Arguments

name - Your prefered name for this index for easier identification.

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

motif - Motif deines what kind of gRNA to search for.

storage_dir - Folder path to the where index will be saved with name linearDB.bin and many prefix files.

prefix_len - Size of the prefix by which off-targets are indexed. Prefix of 8 or larger will be the fastest, however it will also result in large number of files.

skipmer_size - Size of the skipmer as described above. Be carefull when setting this too large!

Examples

# make a temporary directory
tdir = tempname()
mdb_path = joinpath(tdir, "motifDB")
mkpath(mdb_path)

# use ARTEMIS example genome
genome = joinpath(
    vcat(
        splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
        "test", "sample_data", "genome", "semirandom.fa"))

# finally, build a motifDB
build_motifDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 3, ambig_max = 0), 
    mdb_path)

ARTEMIS.search_motifDB — Function

search_motifDB(
    storage_dir::String, 
    guides::Vector{LongDNA{4}}, 
    dist::Int = 4; 
    detail::String = "",
    adjust::Int = 0)

Find all off-targets for guides within distance of dist using motifDB located at storage_dir.

Assumes your guides do not contain PAM, and are all in the same direction as you would order from the lab e.g.:

5' - ...ACGTCATCG NGG - 3'  -> will be input: ...ACGTCATCG
     guide        PAM
    
3' - CCN GGGCATGCT... - 5'  -> will be input: ...AGCATGCCC
     PAM guide

Arguments

dist - Defines maximum levenshtein distance (insertions, deletions, mismatches) for which off-targets are considered.

detail - Path and name for the output file. This search will create intermediate files which will have same name as detail, but with a sequence prefix. Final file will contain all those intermediate files. Leave detail empty if you are only interested in off-target counts returned by the motifDB.

adjust - This will be crutial parameter for tightening second layer of filtering after, the initial prefix alignment. For example, Cas9 off-target within distance 4 (d) might be 20bp long. We skip prefix_len of 7, and are left with 13bp which can be split into 3 skipmers (r) of size 4, 1bp will be left unused. However when searching within distance of 4 and for prefix where initial alignment was of distance 3 (m) and adjustment paramter is 0 (a). We are obliged to find at least k - (d - m + a) which is 3 - (4 - 3 + 0) = 2 this many skimpers inside the off-targets.

Examples

# make a temporary directory
tdir = tempname()
mdb_path = joinpath(tdir, "motifDB")
mkpath(mdb_path)

# use ARTEMIS example genome
coh_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(
    vcat(
        coh_path, 
        "test", "sample_data", "genome", "semirandom.fa"))

# build a motifDB
build_motifDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 3, ambig_max = 0), 
    mdb_path)

# load up example gRNAs
using BioSequences
guides_s = Set(readlines(joinpath(vcat(coh_path, "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, get results!
mdb_res = search_motifDB(mdb_path, guides, 3)

Vantage-Point tree

ARTEMIS.build_treeDB — Function

build_treeDB(
    name::String,
    genomepath::String,
    motif::Motif,
    storage_dir::String,
    prefix_len::Int = 7)

Build a Vantage Point tree DB of offtargets for the given motif, DB groups off-targets by their prefixes, each prefix has its own Vantage Point tree.

Will return a path to the database location, same as storage_dir.

There is an optimization that if the alignment becomes imposible against the prefix we don't search the off-targets grouped inside the prefix. Therefore it is advantageous to select larger prefix than maximum search distance, however in that case number of files also grows.

Arguments

name - Your prefered name for this index for easier identification.

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

motif - Motif deines what kind of gRNA to search for.

storage_dir - Folder path to the where index will be saved with name linearDB.bin and many prefix files.

prefix_len - Size of the prefix by which off-targets are indexed. Prefix of 8 or larger will be the fastest, however it will also result in large number of files.

Examples

# make a temporary directory
tdir = tempname()
tdb_path = joinpath(tdir, "treeDB")
mkpath(tdb_path)

# use ARTEMIS example genome
genome = joinpath(
    vcat(
        splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
        "test", "sample_data", "genome", "semirandom.fa"))

# finally, build a motifDB
build_treeDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 3, ambig_max = 0), 
    tdb_path)

ARTEMIS.inspect_treeDB — Function

inspect_treeDB(
    storage_dir::String; 
    levels::Int = 5, 
    inspect_prefix::String = "")

See small part of the full vantage point tree of the treeDB.

TreeDB can be split based on the distance to the radius (r) into inside (left <= r) and right (outside > r) nodes.

Examples

# make a temporary directory
tdir = tempname()
tdb_path = joinpath(tdir, "treeDB")
mkpath(tdb_path)

# use ARTEMIS example genome
genome = joinpath(
    vcat(
        splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
        "test", "sample_data", "genome", "semirandom.fa"))

# finally, build a motifDB
build_treeDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 3, ambig_max = 0), 
    tdb_path)

# finally, view some part of the database!
inspect_treeDB(tdb_path; inspect_prefix = "CCGTCGC")

ARTEMIS.search_treeDB — Function

search_treeDB(
    storage_dir::String, 
    guides::Vector{LongDNA{4}}, 
    dist::Int = 4; 
    detail::String = "")

Search previously build treeDB database for the off-targets of the guides.

Assumes your guides do not contain PAM, and are all in the same direction as you would order from the lab e.g.:

5' - ...ACGTCATCG NGG - 3'  -> will be input: ...ACGTCATCG
     guide        PAM

3' - CCN GGGCATGCT... - 5'  -> will be input: ...AGCATGCCC
     PAM guide

Arguments

dist - Defines maximum levenshtein distance (insertions, deletions, mismatches) for which off-targets are considered.

detail - Path and name for the output file. This search will create intermediate files which will have same name as detail, but with a sequence prefix. Final file will contain all those intermediate files. Leave detail empty if you are only interested in off-target counts returned by the treeDB.

Examples

# make a temporary directory
tdir = tempname()
tdb_path = joinpath(tdir, "treeDB")
mkpath(tdb_path)

# use ARTEMIS example genome
coh_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(
    vcat(
        coh_path, 
        "test", "sample_data", "genome", "semirandom.fa"))

# build a motifDB
build_treeDB(
    "samirandom", genome, 
    Motif("Cas9"; distance = 3, ambig_max = 0), 
    tdb_path)

# load up example gRNAs
using BioSequences
guides_s = Set(readlines(joinpath(vcat(coh_path, "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, get results!
tdb_res = search_treeDB(tdb_path, guides, 3)