Find all off-targets

Prefix-Suffix partial alignment

ARTEMIS.build_linearDB — Function

build_linearDB(
    name::String,
    genomepath::String,
    motif::Motif,
    storage_dir::String,
    prefix_len::Int = 7)

Prepare linearDB index for future searches using search_linearDB.

Will return a path to the database location, same as storage_dir. When this database is used for the guide off-target scan it is similar to linear in performance, hence the name. There is an optimization that if the alignment becomes impossible against the prefix we don't search the off-targets grouped inside the prefix. Therefore, it is advantageous to select much larger prefix than maximum search distance, however in that case number of files also grows. For example, if interested with searches within distance 4, preferably use prefix length of 7 or 8.

Arguments

name - Your preferred name for this index for easier identification.

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

motif - Motif defines what kind of gRNA to search for.

storage_dir - Folder path to the where index will be saved with name linearDB.bin and many prefix files.

prefix_len - Size of the prefix by which off-targets are indexed. Prefix of 8 or larger will be the fastest, however it will also result in large number of files.

Examples

using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
db_path = joinpath(tdir, "linearDB")
mkpath(db_path)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))

# build a linearDB
build_linearDB(
    "samirandom", genome, 
    Motif("Cas9"), 
    db_path)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(tdir, "linearDB", "results.csv")
search_linearDB(db_path, guides, res_path; distance = 3)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 3)

ARTEMIS.search_linearDB — Function

search_linearDB(
    storage_dir::String, 
    guides::Vector{LongDNA{4}}, 
    output_file::String;
    distance::Int = 4)

Find all off-targets for guides within distance of dist using linearDB located at storage_dir.

Assumes your guides do not contain PAM, and are all in the same direction as you would order from the lab e.g.:

5' - ...ACGTCATCG NGG - 3'  -> will be input: ...ACGTCATCG
     guide        PAM
    
3' - CCN GGGCATGCT... - 5'  -> will be input: ...AGCATGCCC
     PAM guide

Arguments

output_file - Path and name for the output file, this will be comma separated table, therefore .csv extension is preferred. This search will create intermediate files which will have same name as output_file, but with a sequence prefix. Final file will contain all those intermediate files.

distance - Defines maximum levenshtein distance (insertions, deletions, mismatches) for which off-targets are considered.

Examples

using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
db_path = joinpath(tdir, "linearDB")
mkpath(db_path)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))

# build a linearDB
build_linearDB(
    "samirandom", genome, 
    Motif("Cas9"), 
    db_path)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(tdir, "linearDB", "results.csv")
search_linearDB(db_path, guides, res_path; distance = 3)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 3)

ARTEMIS.build_motifDB — Function

build_motifDB(
    name::String,
    genomepath::String,
    motif::Motif,
    storage_dir::String,
    prefix_len::Int = 7;
    skipmer_size::Int = Int(floor(length_noPAM(motif) / (motif.distance + 3))))

Prepare motifDB index for future searches using search_motifDB.

Will return a path to the database location, same as storage_dir. When this database is used for the guide off-target scan it is similar to search_linearDB, however additional filter is applied on top of prefix filtering. Suffixes are used for next filter, similarly to pigeon hole principle - depending on the size of the skipmer skipmer_size. For example, Cas9 off-target within distance 4 (d) might be 20bp long. We skip prefix_len of 7, and are left with 13bp which can be split into 3 skipmer (r) of size 4, 1bp will be left unused. However when searching within distance of 4 and for prefix where initial alignment was of distance 3 (m) and adjustment parameter is 0 (a). We are obliged to find at least r - (d - m + a) which is 3 - (4 - 3 + 0) = 2 this many skipmers inside the off-targets.

There exist also another approach which builds on the idea that it might be more efficient to find at least two kmers of smaller size (named 01*0 seed) rather than one larger kmer (pigeon hole principle). You can use the adjust option for that during search_linearDB step.

Be sure to understand implications of using motifDB as using wrong parameters on skipmer_size might result in leaky filtering in relation to the assumed distance dist and adjustment adjust during search step in search_motifDB.

Arguments

name - Your preferred name for this index for easier identification.

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

motif - Motif defines what kind of gRNA to search for.

storage_dir - Folder path to the where index will be saved and many prefix files.

prefix_len - Size of the prefix by which off-targets are indexed. Prefix of 8 or larger will be the fastest, however it will also result in large number of files.

skipmer_size - Size of the skipmer as described above. Be careful when setting this too large!

Examples

using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
db_path = joinpath(tdir, "motifDB")
mkpath(db_path)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))

# build a motifDB
build_motifDB(
    "samirandom", genome, 
    Motif("Cas9"), 
    db_path)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(tdir, "motifDB", "results.csv")
search_motifDB(db_path, guides, res_path; distance = 3)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 3)

ARTEMIS.search_motifDB — Function

search_motifDB(
    storage_dir::String, 
    guides::Vector{LongDNA{4}}, 
    output_file::String;
    distance::Int = 3,
    adjust::Int = 0)

Find all off-targets for guides within distance of distance using motifDB located at storage_dir.

Assumes your guides do not contain PAM, and are all in the same direction as you would order from the lab e.g.:

5' - ...ACGTCATCG NGG - 3'  -> will be input: ...ACGTCATCG
     guide        PAM
    
3' - CCN GGGCATGCT... - 5'  -> will be input: ...AGCATGCCC
     PAM guide

Arguments

storage_dir - Directory containing motifDB.

guides - a vector of gRNAs without PAM.

output_file - File to which write detailed results. This search will create intermediate files which will have same name as output_file, but with a sequence prefix. Final file will contain all those intermediate files, and other files with sequence prefix will be deleted.

distance - Defines maximum levenshtein distance (insertions, deletions, mismatches) for which off-targets are considered.

adjust - This will be crucial parameter for tightening second layer of filtering after, the initial prefix alignment. For example, Cas9 off-target within distance 4 (d) might be 20bp long. We skip prefix_len of 7, and are left with 13bp which can be split into 3 skipmers (r) of size 4, 1bp will be left unused. However when searching within distance of 4 and for prefix where initial alignment was of distance 3 (m) and adjustment parameter is 0 (a). We are obliged to find at least r - (d - m + a) which is 3 - (4 - 3 + 0) = 2 this many skipmers inside the off-targets.

Examples

using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
db_path = joinpath(tdir, "motifDB")
mkpath(db_path)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))

# build a motifDB
build_motifDB(
    "samirandom", genome, 
    Motif("Cas9"), 
    db_path)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(tdir, "motifDB", "results.csv")
search_motifDB(db_path, guides, res_path; distance = 3)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 3)

Vantage-Point tree

ARTEMIS.build_treeDB — Function

build_treeDB(
    name::String,
    genomepath::String,
    motif::Motif,
    storage_dir::String,
    prefix_len::Int = 7)

Build a Vantage Point tree DB of offtargets for the given motif, DB groups off-targets by their prefixes, each prefix has its own Vantage Point tree.

Will return a path to the database location, same as storage_dir.

There is an optimization that if the alignment becomes impossible against the prefix we don't search the off-targets grouped inside the prefix. Therefore it is advantageous to select larger prefix than maximum search distance, however in that case number of files also grows.

Arguments

name - Your preferred name for this index for easier identification.

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

motif - Motif defines what kind of gRNA to search for.

storage_dir - Folder path to the where index will be saved with name linearDB.bin and many prefix files.

prefix_len - Size of the prefix by which off-targets are indexed. Prefix of 8 or larger will be the fastest, however it will also result in large number of files.

Examples

using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
db_path = joinpath(tdir, "treeDB")
mkpath(db_path)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))

# build a treeDB
build_treeDB(
    "samirandom", genome, 
    Motif("Cas9"), 
    db_path)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(tdir, "treeDB", "results.csv")
search_treeDB(db_path, guides, res_path; distance = 3)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 3)

ARTEMIS.inspect_treeDB — Function

inspect_treeDB(
    storage_dir::String; 
    levels::Int = 5, 
    inspect_prefix::String = "")

See small part of the full vantage point tree of the treeDB.

TreeDB can be split based on the distance to the radius (r) into inside (left <= r) and right (outside > r) nodes.

Examples

using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
db_path = joinpath(tdir, "treeDB")
mkpath(db_path)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))

# build a treeDB
build_treeDB(
    "samirandom", genome, 
    Motif("Cas9"), 
    db_path)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(tdir, "treeDB", "results.csv")
search_treeDB(db_path, guides, res_path; distance = 3)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 3)

# finally, view some part of the database!
inspect_treeDB(db_path; inspect_prefix = "CCGTCGC")

ARTEMIS.search_treeDB — Function

search_treeDB(
    storage_dir::String, 
    guides::Vector{LongDNA{4}}, 
    output_file::String;
    distance::Int = 3)

Search previously build treeDB database for the off-targets of the guides.

Assumes your guides do not contain PAM, and are all in the same direction as you would order from the lab e.g.:

5' - ...ACGTCATCG NGG - 3'  -> will be input: ...ACGTCATCG
     guide        PAM

3' - CCN GGGCATGCT... - 5'  -> will be input: ...AGCATGCCC
     PAM guide

Arguments

dist - Defines maximum levenshtein distance (insertions, deletions, mismatches) for which off-targets are considered.

detail - Path and name for the output file. This search will create intermediate files which will have same name as detail, but with a sequence prefix. Final file will contain all those intermediate files. Leave detail empty if you are only interested in off-target counts returned by the treeDB.

Examples

using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
db_path = joinpath(tdir, "treeDB")
mkpath(db_path)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))

# build a treeDB
build_treeDB(
    "samirandom", genome, 
    Motif("Cas9"), 
    db_path)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(tdir, "treeDB", "results.csv")
search_treeDB(db_path, guides, res_path; distance = 3)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 3)

FM-index

ARTEMIS.build_PathTemplates — Function

build_PathTemplates(len::Int, d::Int; storagepath::String = "", mismatch_only::Bool = false)

Builds up a PathTemplates object. Stores shortened version of all possible paths within each distance d mapped on the graph of all possible alignments of sequence of length len. Then one can use templates_to_sequences_extended or templates_to_sequences and map guide sequence to all possible alignments quickly.

Arguments

len - length of the sequence (e.g. guide - without PAM)

d - Maximal distance on which to build the graph.

storagepath - If not empty "", will save the object under given path.

mismatch_only - Whether to skip insertions/deletions.

build_PathTemplates(motif::Motif; storagepath::String = "", mismatch_only::Bool = false)

Builds up a PathTemplates object. Stores shortened version of all possible paths for given Motif. Afterwards use templates_to_sequences_extended or templates_to_sequences and map guide sequence to all possible alignments quickly.

Arguments

motif - Motif object.

storagepath - If not empty "", will save the object under given path.

mismatch_only - Whether to skip insertions/deletions.

ARTEMIS.build_fmiDB — Function

build_fmiDB(
    genomepath::String,
    storage_dir::String)

Prepare FM-index for future searches.

Arguments

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

storage_dir - Path to the DIRECTORY where index with many files will be saved.

Examples

# prepare libs
using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
fmi_dir = joinpath(tdir, "fmi")
mkpath(fmi_dir)

# use ARTEMIS example genome
genome = joinpath(vcat(splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
    "test", "sample_data", "genome", "semirandom.fa"))

# build a fmiDB!
build_fmiDB(genome, fmi_dir)

ARTEMIS.search_fmiDB — Function

search_fmiDB(
    guides::Vector{LongDNA{4}}, 
    mpt::PathTemplates, 
    motif::Motif, 
    fmidbdir::String,
    output_file::String; 
    distance::Int = 2)

Search FM-index for off-targets using brute-force enumeration method.

Experimental! Proof-of-concept!

This method uses PathTemplates build on top of Motif to enumerate all possible off-target sequences, next, these sequences are found in the genome using FM-index. This method is impractically slow above distance of 2.

Arguments

guides - a vector of gRNAs without PAM.

mpt - PathTemplates object that contains abstraction for all possible alignments

motif - Motif defines what kind of gRNA to search for. Has to be compatible with mpt.

fmidbdir - Path to the folder where FM-index was build using build_fmi.

output_file - Where output will be saved.

distance - Search distance, maximum of 2 is practical.

Examples

# prepare libs
using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
fmi_dir = joinpath(tdir, "fmi")
mkpath(fmi_dir)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))
# build FM-index
build_fmiDB(genome, fmi_dir)
motif = Motif("Cas9"; distance = 1)
mpt = build_PathTemplates(motif)

# prepare output folder
res_dir = joinpath(tdir, "results")
mkpath(res_dir)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(res_dir, "results.csv")
search_fmiDB(guides, mpt, motif, fmi_dir, res_path; distance = 1)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 1)

ARTEMIS.build_pamDB — Function

build_pamDB(
    fmidbdir::String, 
    motif::Motif; 
    storage_path::String = "")

Find locations of all the PAM for a given motif in the genome.

Example of what position we store for traditional Cas9 (NGG, CCN) and Cpf1 (TTTN, NAAA):

NGG  CCN TTTN  NAAA
^      ^    ^  ^

Arguments

fmidbdir - Path to directory with the FM-index build with build_fmiDB.

motif - Motif defines which PAM we will locate in the genome.

storage_path - Path to the DIRECTORY where index will be saved.

Examples

# prepare libs
using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
fmi_dir = joinpath(tdir, "fmi")
mkpath(fmi_dir)

# use ARTEMIS example genome
genome = joinpath(
    vcat(
        splitpath(dirname(pathof(ARTEMIS)))[1:end-1], 
        "test", "sample_data", "genome", "semirandom.fa"))
# build FM-index
build_fmiDB(genome, fmi_dir)

# build a pamDB
pamDB = build_pamDB(fmi_dir, Motif("Cas9"))

ARTEMIS.search_fmiDB_seed — Function

search_fmiDB_seed(
    guides::Vector{LongDNA{4}},
    fmidbdir::String, 
    genomepath::String, 
    pamDB::PAMinFMI,
    output_file::String; 
    distance::Int = 2)

Search FM-index for off-targets using 01*0 seed method. Read more here: publication and pdf.

Experimental! Proof-of-concept!

Arguments

guides - a vector of gRNAs without PAM.

fmidbdir - Path to the folder where FM-index was build using build_fmi.

genomepath - Path to the genome used to build the FM-index.

pamDB - object build with build_pamDB that contains locations of the PAM inside the genome.

output_file - Where output will be saved.

distance - Search distance, maximum of 2 is practical.

Examples

# prepare libs
using ARTEMIS, BioSequences

# make a temporary directory
tdir = tempname()
fmi_dir = joinpath(tdir, "fmi")
mkpath(fmi_dir)

# use ARTEMIS example genome
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))
# build FM-index
build_fmiDB(genome, fmi_dir)

# build a pamDB
motif = Motif("Cas9"; distance = 1)
pamDB = build_pamDB(fmi_dir, motif)

# prepare PathTemplates
mpt = build_PathTemplates(motif)

# prepare output folder
res_dir = joinpath(tdir, "results")
mkpath(res_dir)

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)
    
# finally, make results!
res_path = joinpath(res_dir, "results.csv")
search_fmiDB_seed(guides, fmi_dir, genome, pamDB, res_path; distance = 1)

# load results
using DataFrames, CSV
res = DataFrame(CSV.File(res_path))

# filter results by close proximity
res = filter_overlapping(res, 23)

# summarize results into a table of counts by distance
summary = summarize_offtargets(res, 1)

VCF

ARTEMIS.build_vcfDB — Function

build_vcfDB(
    name::String, 
    genomepath::String, 
    vcfpath::String,
    motif::Motif;
    storage_path::String = "")

Search VcfDB for off-targets for selected guides.

Experimental! Proof-of-concept!

This function does not yet lists all off-targets and this will be added at a later data. Current output is a list of counts of the ambiguous off-targets by distance. Distance is restricted to 0 and 1 at the moment.

Arguments

name - Your preferred name for this index for easier identification.

genomepath - Path to the genome file, it can either be fasta or 2bit file. In case of fasta also prepare fasta index file with ".fai" extension.

vcfpath - Path to the VCF file, it has to be compatible with your genome.

motif - Motif defines what kind of gRNA to search for.

storage_path - Folder path to the where index will be saved.

Examples

# prepare libs
using ARTEMIS, BioSequences

# use ARTEMIS example genome and vcf file
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))
vcf = joinpath(vcat(artemis_path, 
    "test", "sample_data", "artificial.vcf"))

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)

# example VCF file
vcf_db = build_vcfDB(
    "samirandom", genome, vcf,
    Motif("Cas9"; distance = 1, ambig_max = 0))

# search using vcfDB!
vcf_res = search_vcfDB(vcf_db, guides)

ARTEMIS.search_vcfDB — Function

search_vcfDB(
    db::VcfDB,
    guides::Vector{LongDNA{4}})

Search VcfDB for off-targets for selected guides.

Experimental! Proof-of-concept!

This function does not yet lists all off-targets and this will be added at a later data. Current output is a list of counts of the ambiguous off-targets by distance. Distance is restricted to 0 and 1 at the moment.

Arguments

db - object built with build_vcfDB

guides - a vector of gRNAs without PAM.

Examples

# prepare libs
using ARTEMIS, BioSequences

# use ARTEMIS example genome and vcf file
artemis_path = splitpath(dirname(pathof(ARTEMIS)))[1:end-1]
genome = joinpath(vcat(artemis_path, 
    "test", "sample_data", "genome", "semirandom.fa"))
vcf = joinpath(vcat(artemis_path, 
    "test", "sample_data", "artificial.vcf"))

# load up example gRNAs
guides_s = Set(readlines(joinpath(vcat(artemis_path, 
    "test", "sample_data", "crispritz_results", "guides.txt"))))
guides = LongDNA{4}.(guides_s)

# example VCF file
vcf_db = build_vcfDB(
    "samirandom", genome, vcf,
    Motif("Cas9"; distance = 1, ambig_max = 0))

# search using vcfDB!
vcf_res = search_vcfDB(vcf_db, guides)